fix access tests: clean up global variable Image.USE_CFFI_ACCESS after tests do not run tests twice via `test_put` and `test_get`
